+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Make `NetworkEntityMap` public.
+
 ## [0.9.1] - 2023-08-05
 
 ### Changed